requirements.txt
PyPDF2
openai
langchain
langchain-community
faiss-cpu
tqdm
python-dotenv
streamlit
requests